refactor: make sample Deck setup a hook directly
