+++ src/features/deck-import/index.ts
@@ -1,4 +1,4 @@
 export { useDeckImport } from "./hooks/useDeckImport";
-export { useAddSampleDeck } from "./hooks/useAddSampleDeck";
+export { useAddSampleDeck } from "./model/sampleDeck";
 export { downloadSampleCsv, SAMPLE_CSV_TEXT } from "./sampleCsv";
 export { DeckImportView } from "./ui/DeckImportView";
