+++ src/features/deck-import/model/sampleDeck.spec.ts
@@ -1,54 +1,74 @@
-import type { CardMutation } from "@/entities/card";
-import type { DeckCreateInput } from "@/entities/deck";
+import type { Card, CardMutation } from "@/entities/card";
+import type { Deck, DeckCreateInput } from "@/entities/deck";
 
+import { renderHook, waitFor } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { createDeck } from "@/test/factories";
 
 const mocks = vi.hoisted(() => ({
+  uid: "uid-a",
+  cards: [] as Card[],
+  decks: [] as Deck[],
+  createDeck: vi.fn<(_uid: string, _deck: DeckCreateInput) => Promise<unknown>>(),
+  generateCardId: vi.fn(() => crypto.randomUUID()),
   mutateCards: vi.fn<(_uid: string, _mutations: CardMutation[]) => Promise<void>>(),
 }));
 
 vi.mock("@/shared/firebase", () => ({ auth: {}, db: {} }));
+vi.mock("@/entities/auth", () => ({ useAuthUid: () => mocks.uid }));
 vi.mock("@/entities/card", async (importOriginal) => {
   const actual = await importOriginal<typeof import("@/entities/card")>();
-  return { ...actual, mutateCards: mocks.mutateCards };
+  return {
+    ...actual,
+    generateCardId: mocks.generateCardId,
+    mutateCards: mocks.mutateCards,
+    useCards: () => mocks.cards,
+  };
+});
+vi.mock("@/entities/deck", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("@/entities/deck")>();
+  return { ...actual, createDeck: mocks.createDeck, useDecks: () => mocks.decks };
 });
 
-import { addSampleDeck } from "./sampleDeck";
+import { useAddSampleDeck } from "./sampleDeck";
 
-describe("addSampleDeck", () => {
-  const createRemoteDeck = vi.fn<(_uid: string, _deck: DeckCreateInput) => Promise<void>>();
-  const generateCardId = vi.fn(() => crypto.randomUUID());
-
+describe("useAddSampleDeck", () => {
   beforeEach(() => {
     vi.clearAllMocks();
-    createRemoteDeck.mockResolvedValue(undefined);
+    mocks.uid = "uid-a";
+    mocks.cards = [];
+    mocks.decks = [];
+    mocks.createDeck.mockResolvedValue(undefined);
     mocks.mutateCards.mockResolvedValue(undefined);
   });
 
-  it("does not add a sample for a signed-out user", async () => {
-    await addSampleDeck("", { cards: [], createDeck: createRemoteDeck, decks: [], generateCardId });
+  it("does not add a sample for a signed-out user", () => {
+    mocks.uid = "";
 
-    expect(createRemoteDeck).not.toHaveBeenCalled();
+    renderHook(useAddSampleDeck);
+
+    expect(mocks.createDeck).not.toHaveBeenCalled();
     expect(mocks.mutateCards).not.toHaveBeenCalled();
   });
 
-  it("does not add a sample when the user already has a Deck", async () => {
-    const deck = createDeck({ uid: "uid-a" });
+  it("does not add a sample when the user already has a Deck", () => {
+    mocks.decks = [createDeck({ uid: "uid-a" })];
 
-    await addSampleDeck("uid-a", { cards: [], createDeck: createRemoteDeck, decks: [deck], generateCardId });
+    renderHook(useAddSampleDeck);
 
-    expect(createRemoteDeck).not.toHaveBeenCalled();
+    expect(mocks.createDeck).not.toHaveBeenCalled();
     expect(mocks.mutateCards).not.toHaveBeenCalled();
   });
 
   it("adds the sample when the user has no Decks", async () => {
-    await addSampleDeck("uid-a", { cards: [], createDeck: createRemoteDeck, decks: [], generateCardId });
+    renderHook(useAddSampleDeck);
 
-    expect(createRemoteDeck).toHaveBeenCalledWith(
-      "uid-a",
-      expect.objectContaining({ id: "sample-v1-uid-a", name: "Sample Deck", uid: "uid-a" })
+    await waitFor(() =>
+      expect(mocks.createDeck).toHaveBeenCalledWith(
+        "uid-a",
+        expect.objectContaining({ id: "sample-v1-uid-a", name: "Sample Deck", uid: "uid-a" })
+      )
     );
     expect(mocks.mutateCards).toHaveBeenCalledOnce();
   });
