+++ src/features/deck-import/model/sampleDeck.ts
@@ -1,7 +1,11 @@
 import type { Card } from "@/entities/card";
-import type { Deck, DeckCreateInput, DeckId } from "@/entities/deck";
+import type { Deck, DeckId } from "@/entities/deck";
 
-import { mutateCards } from "@/entities/card";
+import { useEffect } from "react";
+
+import { useAuthUid } from "@/entities/auth";
+import { generateCardId, mutateCards, useCards } from "@/entities/card";
+import { createDeck, useDecks } from "@/entities/deck";
 import sampleCards from "../../../../sample/build/output.json";
 import { executePreparedDeckImport, prepareDeckImport } from "./deckImportExecution";
 
@@ -10,14 +14,13 @@ const SAMPLE_VERSION = 1;
 
 const sampleDeckId = (uid: string): DeckId => `sample-v${SAMPLE_VERSION}-${uid}`;
 
-export interface SampleDeckOptions {
+export interface SampleDeckPreparationOptions {
   cards: Card[];
-  createDeck: (uid: string, deck: DeckCreateInput) => Promise<unknown>;
   decks: Deck[];
   generateCardId: () => string;
 }
 
-export const prepareSampleDeck = (uid: string, options: Omit<SampleDeckOptions, "createDeck">) =>
+export const prepareSampleDeck = (uid: string, options: SampleDeckPreparationOptions) =>
   prepareDeckImport(
     {
       name: SAMPLE_DECK_NAME,
@@ -27,16 +30,23 @@ export const prepareSampleDeck = (uid: string, options: Omit<SampleDeckOptions, 
     { uid, ...options }
   );
 
-export const addSampleDeck = (uid: string, options: SampleDeckOptions) => {
-  if (uid === "" || options.decks.length > 0) return Promise.resolve();
+export const useAddSampleDeck = () => {
+  const uid = useAuthUid();
+  const cards = useCards();
+  const decks = useDecks();
 
-  return executePreparedDeckImport(prepareSampleDeck(uid, options), {
-    uid,
-    createDeck: (deck) => {
-      // Samples remain account-synced even when CSV imports support a local destination.
-      if (deck.localMode === true) throw new Error("The sample Deck cannot use local storage");
-      return options.createDeck(uid, deck);
-    },
-    mutateCards: (mutations) => mutateCards(uid, mutations),
-  });
+  useEffect(() => {
+    if (uid === "" || decks.length > 0) return;
+
+    // Bootstrap is opportunistic; subscription changes provide another attempt without blocking the Deck list.
+    void executePreparedDeckImport(prepareSampleDeck(uid, { cards, decks, generateCardId }), {
+      uid,
+      createDeck: (deck) => {
+        // Samples remain account-synced even when CSV imports support a local destination.
+        if (deck.localMode === true) throw new Error("The sample Deck cannot use local storage");
+        return createDeck(uid, deck);
+      },
+      mutateCards: (mutations) => mutateCards(uid, mutations),
+    }).catch(() => undefined);
+  }, [cards, decks, uid]);
 };
